experiments/requirements.txt
datasets==2.8.0
nltk==3.5
numpy==1.21.2
pandas==1.3.0
pytorch==1.11.0
rouge-score==0.1.2
tqdm==4.64.0
transformers==4.25.1